feat(trip-planner): read a booking confirmation into trip items
Adds "Read a booking PDF" to the trip menu. It takes a flight or hotel
confirmation - a PDF or pasted text - and offers what it found as ordinary
proposal cards, so accepting one runs the same validation, undo and save
path as an assistant suggestion.

It is deterministic, not AI. Airport codes resolve against the bundled
table; dates, times, booking reference and price are read by pattern. The
extractor moves into trip-logic.js with the rest of the pure logic, and
its 45 tests move into the main suite.

THE FILE NEVER LEAVES THE DEVICE AND IS NEVER STORED. It is read into
memory, turned into text and dropped - nothing reaches the documents
pocket, nothing is uploaded, no model is called. PDF text extraction is
dependency-free (about forty lines over DecompressionStream) rather than
1.7 MB of vendored pdf.js; a PDF it cannot read, including any scan, is
refused with a pointer to the paste box instead of guessed at.

Every reading shows the LINE it came from and a confidence, because a
wrong read has to be visible rather than merely plausible in a form
field.

New concept: transcription provenance. sanitizeActionFields refuses a
confirmation code and a booked status from a MODEL, because a model can
only invent one and a plausible fake is worse at a check-in desk than a
blank. Copying one off the traveller's own confirmation is the opposite
situation, so an action marked source:'document' may carry both, and a
transcribed price counts as real cost rather than an estimate. It unlocks
exactly those fields and relaxes nothing else: a negative cost is still
dropped, a bad type is still rejected, cancelled still never survives.
The default is unchanged, so every existing assistant path behaves
exactly as before.
