+++ package.json
@@ -24,9 +24,7 @@
     "enrich:rising-shows": "node apps/rising-shows/scripts/enrich-tmdb.js",
     "enrich:rising-shows:providers": "node apps/rising-shows/scripts/enrich-providers.js",
     "export:rising-shows": "node apps/rising-shows/scripts/export-integrations.js",
-    "watch-next": "node apps/rising-shows/scripts/watch-next.js",
-    "proto:extract": "node apps/trip-planner/scripts/proto/extract-cli.mjs",
-    "test:tp-extract-proto": "node --test apps/trip-planner/scripts/proto/"
+    "watch-next": "node apps/rising-shows/scripts/watch-next.js"
   },
   "dependencies": {
     "@netlify/blobs": "^8.1.0"
